[ENH,WIP] RegistrationReportCapable interfaces
Creates an abstract mixin for registration interfaces, where two images
(the reference and the moving) are to be presented with the flickering
animation.

This first commit covers the following:

- Split the visualization utilities from common/report.py to viz/utils.py
- Some house-keeping on the base class:
  - Change the main report output name to ```out_report``` to be agnostic to the output format
  - Fix the super classes of the ReportCapableInput/OutputSpecs to be traited (they won't work otherwise)
  - Simplify _run_interface: avoid using a try...except to wrap the super()._run_interface (less error tolerant)
  - Add logging
  - Pass exit code to _generate_error_report
